Add files via upload
A change from planned or expected behavior or a sequence of events that endangers or negatively affects people, property, or the environment is defined as an emergency. On the other hand, it can be thought of as a midpoint between collision detection and prevention. This paper presents the results of a major analysis of collision (automobile) emergency alert conditions. In this study, the authors have investigated modern Internet of Things (IoT) and VANET (Vehicular Ad hoc Networks) technologies and developed a collection of modern and specialized techniques, as well as their characteristics. It has sensors that detect unbalanced vehicles and provide a warning to the microcontroller if a collision occurs. The technology has been designed in such a way that it would not only detect but also prevent a collision in any situation. Additionally, the technique can be implemented in such a way that vehicles are alerted of possible closing barriers. Vehicle-to-vehicle communication (V2V) has a huge impact since it allows vehicles to communicate with one another while in close proximity, and the buzzer, together with the LEDs, serves as a safety feature. The system's primary goal is to carry out the microcontroller functions in every environment, moreover the concept refers to detect and prevent the collision in a foggy weather and at night. Wireless technology, along with the Internet of Things (IoT), has resulted in a vast and significant expansion of VANETs infrastructure in recent years (Vehicular Ad hoc Networks). The Internet of Things (IoT) and the Vehicular Ad-Hoc Network (VANET) have now been merged as the fundamental and central components of today's Intelligent Transportation System (ITS). As a result, we'd like to create a system that notifies the driver when a situation is potentially dangerous and quickly shares that area in the event of a collision. If any type of collision occurs, the system will provide emergency assistance such as hospitals, police stations, and insurance companies. A message will be sent to the registered family member's phone number as well. Furthermore, while the procedure of obtaining the insurance may be longer for certain people, others may avoid the law after being involved in severe collisions, making it difficult for authorities to discriminate between criminal and non-criminal evidence.